Fix redeclaration errors
